backend/requirements.txt
asgiref==3.11.0
bcrypt==5.0.0
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.5
whitenoise==6.11.0
